Example file contents
