+++ webui/extensionlib/extensionstates.py
@@ -1,4 +1,14 @@
-from webui.extensionlib.extensionInfo import ExtensionInfo
+import shlex
+import subprocess
+
+from setup_tools import os
+
+
+def git_ready():
+    cmd = 'git -v'
+    cmd = cmd if os.is_windows() else shlex.split(cmd)
+    result = subprocess.run(cmd).returncode
+    return result == 0
 
 
 class ExtensionState:
